misc: mtk-connectivity: Set proper wlan connectivity driver specifications
MTK WiFi drivers are userspace kernel modules, that have appended flags
during compile time from the board tree. These are inherited from the
device tree and hence do not work as intended when built into kernel
as the specifications are not met.

Certain flags are needed to get the driver working when built in kernel.
Namely:
 -> MTK_COMBO_CHIP
 -> WLAN_CHIP_ID
 -> WIFI_IP_SET
 -> WIFI_ECO_VER

To setup these flags properly, firmware binaries are assessed that are
located in /vendor/firmware/ in the stock OS/firmware.

The binaries needed are
 -> WIFI_RAM_CODE_socXX_XX_X.bin
 -> socXX_ram_wifi_XX_X_hdr.bin

For example, MT6853T (Dimensity 800U) has these
 -> WIFI_RAM_CODE_soc2_0_3c_1.bin
 -> soc2_0_ram_wifi_3c_1_hdr.bin

Binaries need to be split for exact properties.
 - WIFI_RAM_CODE_soc2_0_3c_1.bin
                     |  |  |
                     |  |   -- "1" indicates WIFI_ECO_VER
                     |   -- "3" indicates WIFI_IP_SET, the letter "c"
                     |      indicates WIFI_FLAVOR, unneeded.
                      -- 2_0 indicates the WiFi branch.

 Similar information can be extracted from:
 - soc2_0_ram_wifi_3c_1_hdr.bin
       |           |  |
       |           |   -- WIFI_ECO_VER
       |            -- WIFI_IP_SET
        -- WiFi branch

To figure out MTK_COMBO_CHIP using the WiFi branch, MTK has divided
it on the basis of chips. Use this as reference:
For WiFi branch values of:
  - 1_0 -> MTK_COMBO_CHIP=CONNAC
  - 2_0 -> MTK_COMBO_CHIP=CONNAC2X2
  - 3_0 -> MTK_COMBO_CHIP=CONNAC2X2_SOC3_0
 (Note: This is subject to change when MTK Releases newer chips. This is
  based on MTK Android 10 specifications.)

The WLAN_CHIP_ID flag is usually the same as the chipset code, i.e.,
6853 for MT6853. Note that MTK WiFi drivers are listed with generic
6765 code, which may cause compile errors and lead to wifi not work,
unless your chipset is mt6765.

So for MT6853, the values are:
 -> MTK_COMBO_CHIP=CONNAC2X2
 -> WIFI_IP_SET=3
 -> WIFI_ECO_VER=1

The flag MTK_ANDROID_EMI is subject to change from chipset to chipset.
Only trial and error can help with it. But it can be enabled by default.

Additionally, Adaptor options maybe needed.

Test:
- Apply according to the details provided above.
- Boot, check wifi.

Result:
- Wifi works as intended.

Signed-off-by: Vaisakh Murali <mvaisakh@statixos.com>